Removing metals from loot tables.
Removed lead, tin, nickel and platinum from the loot tables.

I know we never originally discussed platinum but I just discovered that platinum is exclusively used for making enderium, which is exclusively used to make vibrant alloy. If we re-introduce it I'd like it to be a product of industry.