Railcraft kits made at the engineering bench
I've done this to make it so that all railcraft track kits are made of the same cheap parts.

I did this because in our current version of the pack railcraft is completely abandoned and forgotten content, and I think that's a shame, so to encourage its use I am going to make it exceptionally easy and convenient for players to start laying track and using minecarts by making the mod very cheap and easy to get in to.